+++ Model/t1BuildInc/CH4Moderator.h
@@ -98,7 +98,7 @@ class CH4Moderator : public attachSystem::ContainedComp,
 
   virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual int getLayerSurf(const size_t,const long int) const;
-  virtual std::string getLayerString(const size_t,const size_t) const;
+  virtual std::string getLayerString(const size_t,const long int) const;
 
   void createAll(Simulation&,const attachSystem::FixedComp&);
