+++ System/constructInc/CylMod.h
@@ -77,7 +77,7 @@ class CylMod : public constructSystem::ModBase
   virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual int getLayerSurf(const size_t,const long int) const;
   virtual int getCommonSurf(const size_t) const;
-  virtual std::string getLayerString(const size_t,const size_t) const;
+  virtual std::string getLayerString(const size_t,const long int) const;
 
   void createAll(Simulation&,const attachSystem::FixedComp&,
 		 const attachSystem::FixedComp*,
